v2-pipeline/pyproject.toml
[tool.poetry]
name = "v2-pipeline"
version = "0.1.0"
description = ""
authors = ["PRANJAL KAR <74347116+pranjalkar99@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.12"
annotated-types = "^0.6.0"
anyio = "^4.2.0"
aspose-words = "^23.12.0"
beautifulsoup4 = "^4.12.2"
bs4 = "^0.0.1"
cachetools = "^5.3.2"
certifi = "^2023.11.17"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
dnspython = "^2.4.2"
email-validator = "^2.1.0.post1"
exceptiongroup = "^1.2.0"
fastapi = "^0.108.0"
google-api-core = "^2.15.0"
google-api-python-client = "^2.112.0"
google-auth = "^2.26.1"
google-auth-httplib2 = "^0.2.0"
google-cloud-core = "^2.4.1"
google-cloud-storage = "^2.14.0"
google-crc32c = "^1.5.0"
google-resumable-media = "^2.7.0"
googleapis-common-protos = "^1.62.0"
h11 = "^0.14.0"
httpcore = "^1.0.2"
httplib2 = "^0.22.0"
httptools = "^0.6.1"
httpx = "^0.26.0"
idna = "^3.6"
iniconfig = "^2.0.0"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
markupsafe = "^2.1.3"
orjson = "^3.9.10"
packaging = "^23.2"
pluggy = "^1.3.0"
protobuf = "^4.25.1"
pyasn1 = "^0.5.1"
pyasn1-modules = "^0.3.0"
pydantic = "^2.5.3"
pydantic-core = "^2.14.6"
pymongo = "^4.6.1"
pyparsing = "^3.1.1"
pypdf2 = "^3.0.1"
pytest = "^7.4.4"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"
requests = "^2.31.0"
rsa = "^4.9"
six = "^1.16.0"
sniffio = "^1.3.0"
soupsieve = "^2.5"
starlette = ">=0.29.0"
tomli = "^2.0.1"
tqdm = "^4.66.1"
typing-extensions = "^4.9.0"
ujson = "^5.9.0"
uritemplate = "^4.1.1"
urllib3 = "^2.1.0"
uvicorn = "^0.25.0"
watchfiles = "^0.21.0"
websockets = "^12.0"
llama-index = "^0.9.26"
celery = "^5.3.6"
redis = "^5.0.1"
pypdf = "^3.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
